wip: refactoring ui spending notifications
