+++ dashboard/src/features/orgs/components/billing/components/BillingEstimate/components/BillingDetails/index.ts
@@ -1,0 +1,1 @@
+export { default as BillingDetails } from './BillingDetails';
