+++ dashboard/src/features/orgs/components/billing/components/BillingEstimate/components/Estimate/index.ts
@@ -1,0 +1,1 @@
+export { default as Estimate } from './Estimate';
