+++ dashboard/src/features/orgs/components/billing/components/BillingEstimate/index.ts
@@ -1,0 +1,1 @@
+export { default as BillingEstimate } from './BillingEstimate';
