--- dashboard/src/features/orgs/components/billing/components/Usage/index.ts
@@ -1,1 +1,0 @@
-export { default as Usage } from './Usage';
