+++ dashboard/src/features/orgs/components/billing/utils/getBillingCycle.ts
@@ -18,11 +18,13 @@ export const getBillingCycleInfo = () => {
     (now.getTime() - startOfMonth.getTime()) / (1000 * 60 * 60 * 24) + 1;
 
   const progress = (daysPassed / totalDays) * 100;
+  const daysLeft = Math.max(Math.ceil(totalDays - daysPassed), 0);
 
   return {
     billingCycleStart,
     billingCycleEnd,
     billingCycleRange: `${billingCycleStart} - ${billingCycleEnd}`,
     progress: Math.min(Math.max(progress, 0), 100), // Ensure the value is between 0 and 100
+    daysLeft,
   };
 };
