+++ dashboard/src/pages/orgs/[orgSlug]/billing.tsx
@@ -1,5 +1,5 @@
+import { BillingEstimate } from '@/features/orgs/components/billing/components/BillingEstimate';
 import { SubscriptionPlan } from '@/features/orgs/components/billing/components/SubscriptionPlan';
-import { Usage } from '@/features/orgs/components/billing/components/Usage';
 import { ProjectLayout } from '@/features/orgs/layout/ProjectLayout';
 import { useCurrentOrg } from '@/features/orgs/projects/hooks/useCurrentOrg';
 import type { ReactElement } from 'react';
@@ -9,7 +9,7 @@ export default function OrgBilling() {
   return (
     <div className="flex h-full flex-col gap-4 overflow-auto bg-accent p-4">
       <SubscriptionPlan />
-      {!isFree && <Usage />}
+      {!isFree && <BillingEstimate />}
     </div>
   );
 }
